+++ packages/tsurlfilter/vitest.config.ts
@@ -2,6 +2,9 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
     test: {
+        setupFiles: [
+            './test/setup/custom-matchers/index.ts',
+        ],
         watch: false,
     },
 });